+++ aiohttp_client_cache/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.7.0'
+__version__ = '0.7.1'
 
 # flake8: noqa: F401, F403
 try: