+++ HISTORY.md
@@ -4,7 +4,8 @@
 [See all issues & PRs for v0.7](https://github.com/JWCook/aiohttp-client-cache/milestone/6?closed=1)
 * Support manually saving a response to the cache with `CachedSession.cache.save_response()`
 * Add compatibility with aioboto3 0.9+
-* Add `CachedSession` properties derived from headers:
+* Migrate to redis-py 4.2+ (merged with aioredis)
+* Add missing `CachedSession` properties derived from headers:
   * `charset`
   * `content_length`
   * `content_type`